Trait ark_relations::r1cs::Field

pub trait Field: 'static + ToBytes + FromBytes + Copy + Clone + Debug + Display + Default + Send + Sync + Eq + Zero<Output = Self, Output = Self> + One<Output = Self, Output = Self> + Ord + Neg<Output = Self> + UniformRand + Zeroize + Hash + CanonicalSerialize + CanonicalSerializeWithFlags + CanonicalDeserialize + CanonicalDeserializeWithFlags + Add<Self> + Sub<Self, Output = Self, Output = Self> + Mul<Self> + Div<Self, Output = Self, Output = Self> + AddAssign<Self> + SubAssign<Self> + MulAssign<Self> + DivAssign<Self> + for<'a> Add<&'a Self> + for<'a> Sub<&'a Self> + for<'a> Mul<&'a Self> + for<'a> Div<&'a Self> + for<'a> AddAssign<&'a Self> + for<'a> SubAssign<&'a Self> + for<'a> MulAssign<&'a Self> + for<'a> DivAssign<&'a Self> + Sum<Self> + for<'a> Sum<&'a Self> + Product<Self> + for<'a> Product<&'a Self> + From<u128> + From<u64> + From<u32> + From<u16> + From<u8> + From<bool> {
    type BasePrimeField: PrimeField;

    fn extension_degree() -> u64;

    fn from_base_prime_field_elems(
        elems: &[Self::BasePrimeField]
    ) -> Option<Self>;

    fn double(&self) -> Self;

    fn double_in_place(&mut self) -> &mut Self;

    fn from_random_bytes_with_flags<F>(bytes: &[u8]) -> Option<(Self, F)>
    where
        F: Flags;

    fn square(&self) -> Self;

    fn square_in_place(&mut self) -> &mut Self;

    fn inverse(&self) -> Option<Self>;

    fn inverse_in_place(&mut self) -> Option<&mut Self>;

    fn frobenius_map(&mut self, power: usize);

    fn characteristic() -> &'static [u64] { ... }

    fn from_random_bytes(bytes: &[u8]) -> Option<Self> { ... }

    fn pow<S>(&self, exp: S) -> Self
    where
        S: AsRef<[u64]>,
    { ... }

    fn pow_with_table<S>(powers_of_2: &[Self], exp: S) -> Option<Self>
    where
        S: AsRef<[u64]>,
    { ... }

}

The interface for a generic field.

Associated Types

type BasePrimeField: PrimeField

Required methods

fn extension_degree() -> u64

Returns the extension degree of this field with respect to Self::BasePrimeField.

fn from_base_prime_field_elems(elems: &[Self::BasePrimeField]) -> Option<Self>

Convert a slice of base prime field elements into a field element. If the slice length != Self::extension_degree(), must return None.

fn double(&self) -> Self

Returns self + self.

fn double_in_place(&mut self) -> &mut Self

Doubles self in place.

fn from_random_bytes_with_flags<F>(bytes: &[u8]) -> Option<(Self, F)> where
F: Flags,

Returns a field element with an extra sign bit used for group parsing if the set of bytes forms a valid field element, otherwise returns None. This function is primarily intended for sampling random field elements from a hash-function or RNG output.

fn square(&self) -> Self

Returns self * self.

fn square_in_place(&mut self) -> &mut Self

Squares self in place.

fn inverse(&self) -> Option<Self>

Computes the multiplicative inverse of self if self is nonzero.

fn inverse_in_place(&mut self) -> Option<&mut Self>

fn frobenius_map(&mut self, power: usize)

Exponentiates this element by a power of the base prime modulus via the Frobenius automorphism.

Provided methods

fn characteristic() -> &'static [u64]

Returns the characteristic of the field, in little-endian representation.

fn from_random_bytes(bytes: &[u8]) -> Option<Self>

Returns a field element if the set of bytes forms a valid field element, otherwise returns None. This function is primarily intended for sampling random field elements from a hash-function or RNG output.

fn pow<S>(&self, exp: S) -> Self where
S: AsRef<[u64]>,

Exponentiates this element by a number represented with u64 limbs, least significant limb first.

fn pow_with_table<S>(powers_of_2: &[Self], exp: S) -> Option<Self> where
S: AsRef<[u64]>,

Exponentiates a field element f by a number represented with u64 limbs, using a precomputed table containing as many powers of 2 of f as the 1 + the floor of log2 of the exponent exp, starting from the 1st power. That is, powers_of_2 should equal &[p, p^2, p^4, ..., p^(2^n)] when exp has at most n bits.

This returns None when a power is missing from the table.