ark_relations::r1cs

Enum ConstraintSystemRef

pub enum ConstraintSystemRef<F: Field> {
    None,
    CS(Rc<RefCell<ConstraintSystem<F>>>),
}

A shared reference to a constraint system that can be stored in high level variables.

Variants

None

Represents the case where we donโt need to allocate variables or enforce constraints. Encountered when operating over constant values.

CS(Rc<RefCell<ConstraintSystem<F>>>)

Represents the case where we do allocate variables or enforce constraints.

Implementations

impl<F: Field> ConstraintSystemRef<F>

pub fn or(self, other: Self) -> Self

Returns self if !self.is_none(), otherwise returns other.

pub fn is_none(&self) -> bool

Returns true is self == ConstraintSystemRef::None.

pub fn new(inner: ConstraintSystem<F>) -> Self

Construct a ConstraintSystemRef from a ConstraintSystem.

pub fn into_inner(self) -> Option<ConstraintSystem<F>>

Consumes self to return the inner ConstraintSystem<F>. Returns None if Self::CS is None or if any other references to Self::CS exist.

pub fn borrow(&self) -> Option<Ref<'_, ConstraintSystem<F>>>

Obtain an immutable reference to the underlying ConstraintSystem.

Panics

This method panics if self is already mutably borrowed.

pub fn borrow_mut(&self) -> Option<RefMut<'_, ConstraintSystem<F>>>

Obtain a mutable reference to the underlying ConstraintSystem.

Panics

This method panics if self is already mutably borrowed.

pub fn set_mode(&self, mode: SynthesisMode)

Set self.mode to mode.

pub fn is_in_setup_mode(&self) -> bool

Check whether self.mode == SynthesisMode::Setup.

pub fn num_constraints(&self) -> usize

Returns the number of constraints.

pub fn num_instance_variables(&self) -> usize

Returns the number of instance variables.

pub fn num_witness_variables(&self) -> usize

Returns the number of witness variables.

pub fn optimization_goal(&self) -> OptimizationGoal

Check whether this constraint system aims to optimize weight, number of constraints, or neither.

pub fn set_optimization_goal(&self, goal: OptimizationGoal)

Specify whether this constraint system should aim to optimize weight, number of constraints, or neither.

pub fn should_construct_matrices(&self) -> bool

Check whether or not self will construct matrices.

pub fn new_input_variable<Func>(&self, f: Func) -> Result<Variable>
where Func: FnOnce() -> Result<F>,

Obtain a variable representing a new public instance input.

pub fn new_witness_variable<Func>(&self, f: Func) -> Result<Variable>
where Func: FnOnce() -> Result<F>,

Obtain a variable representing a new private witness input.

pub fn new_lc(&self, lc: LinearCombination<F>) -> Result<Variable>

Obtain a variable representing a linear combination.

pub fn enforce_constraint( &self, a: LinearCombination<F>, b: LinearCombination<F>, c: LinearCombination<F>, ) -> Result<()>

Enforce a R1CS constraint with the name name.

pub fn inline_all_lcs(&self)

Naively inlines symbolic linear combinations into the linear combinations that use them.

Useful for standard pairing-based SNARKs where addition gates are cheap. For example, in the SNARKs such as [Groth16] and [Groth-Maller17], addition gates do not contribute to the size of the multi-scalar multiplication, which is the dominating cost.

pub fn finalize(&self)

Finalize the constraint system (either by outlining or inlining, if an optimization goal is set).

pub fn to_matrices(&self) -> Option<ConstraintMatrices<F>>

This step must be called after constraint generation has completed, and after all symbolic LCs have been inlined into the places that they are used.

pub fn is_satisfied(&self) -> Result<bool>

If self is satisfied, outputs Ok(true). If self is unsatisfied, outputs Ok(false). If self.is_in_setup_mode() or if self == None, outputs Err(()).

pub fn which_is_unsatisfied(&self) -> Result<Option<String>>

If self is satisfied, outputs Ok(None). If self is unsatisfied, outputs Some(i), where i is the index of the first unsatisfied constraint. If self.is_in_setup_mode() or self == None, outputs Err(()).

pub fn assigned_value(&self, v: Variable) -> Option<F>

Obtain the assignment corresponding to the Variable v.

pub fn constraint_names(&self) -> Option<Vec<String>>

Get trace information about all constraints in the system